doc_processor/pyproject.toml
[project]
name = "invoice-processor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.5.0",
    "pydantic",
    "python-dotenv",
    "google-cloud-bigquery",
]